Use a base controller for API
