+++ app/controllers/api/ping_controller.rb
@@ -1,12 +1,14 @@
-class API::PingController < APIController
-  skip_before_action :authenticate_user!
+module API
+  class PingController < BaseController
+    skip_before_action :authenticate_user!
 
-  def index
-    render json: {
-      status: {
-        website: true,
-        database: true
-      }
-    }, status: 200
+    def index
+      render json: {
+        status: {
+          website: true,
+          database: true
+        }
+      }, status: 200
+    end
   end
 end
