+++ app/controllers/api/validate_token_controller.rb
@@ -1,9 +1,11 @@
-class API::ValidateTokenController < APIController
-  def index
-    render json: {
-      status: {
-        token: 'valid',
-      }
-    }, status: 200
+module API
+  class ValidateTokenController < BaseController
+    def index
+      render json: {
+        status: {
+          token: 'valid',
+        }
+      }, status: 200
+    end
   end
 end
