+++ app/controllers/api/webhooks/analyzers_controller.rb
@@ -1,19 +1,21 @@
-class API::Webhooks::AnalyzersController < API::WebhooksController
-  before_action :verify_github_webhook
+module API
+  module Webhooks
+    class AnalyzersController < WebhooksController
+      def create
+        repo = params[:repository][:name]
+        track_slug = repo.gsub(/-analyzer$/, '').to_sym
 
-  def create
-    repo = params[:repository][:name]
-    track_slug = repo.gsub(/-analyzer$/, '').to_sym
+        return unless Exercism::ANALYZERS.include?(track_slug)
 
-    return unless Exercism::ANALYZERS.include?(track_slug)
+        tag = params[:release][:tag_name]
 
-    tag = params[:release][:tag_name]
+        PubSub::PublishMessage.(:analyzer_ready_to_build, {
+          track_slug: track_slug,
+          tag: tag,
+        })
 
-    PubSub::PublishMessage.(:analyzer_ready_to_build, {
-      track_slug: track_slug,
-      tag: tag,
-    })
-
-    head :ok
+        head :ok
+      end
+    end
   end
 end
