+++ app/controllers/api/webhooks/repo_updates_controller.rb
@@ -1,27 +1,31 @@
-class API::Webhooks::RepoUpdatesController < API::WebhooksController
-  before_action :verify_github_webhook
+module API
+  module Webhooks
+    class RepoUpdatesController < WebhooksController
+      before_action :verify_github_webhook
 
-  MASTER_REF = "refs/heads/master"
+      MASTER_REF = "refs/heads/master"
 
-  def create
-    return unless pushed_to_master?
+      def create
+        return unless pushed_to_master?
 
-    repo_update = RepoUpdate.new(slug: slug)
+        repo_update = RepoUpdate.new(slug: slug)
 
-    render json: {} if repo_update.save
-  end
+        render json: {} if repo_update.save
+      end
 
-  private
+      private
 
-  def payload_params
-    params.permit(:ref, { repository: :name })
-  end
+      def payload_params
+        params.permit(:ref, { repository: :name })
+      end
 
-  def pushed_to_master?
-    payload_params[:ref].eql? MASTER_REF
-  end
+      def pushed_to_master?
+        payload_params[:ref].eql? MASTER_REF
+      end
 
-  def slug
-    payload_params[:repository][:name]
+      def slug
+        payload_params[:repository][:name]
+      end
+    end
   end
 end
